fixes from tom's review
